Add files via upload
These are the complete set of files that will aid to apply the approximate periodic boundary condition.
The methodology was proposed in 10.2140/jomms.2017.12.471 by Schneider, Konrad , Klusemann, Benjamin , Bargmann, Swantje. 

The following code will help one to apply this method with in the framework on ABAQUS. 

For any quires, ping: krishna.klume@gmail.com

For example, run code Model_E_mesh_Conv.py to apply the APBC on the model with circular fiber inside  a matrix.